requirements.txt
pytest-cov
pytest-mock
requests-mock
pytruth
hypothesis
pyahocorasick
genbadge[all]
